requirements.txt
alembic==1.14.1
bleach==6.0.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.0.0
Flask==2.3.3
Flask-Admin==1.6.1
Flask-AdminLTE3==1.0.9
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Flask-CKEditor==0.4.6
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
Markdown==3.4.4
MarkupSafe==3.0.2
Pillow==10.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==2.3.7
WTForms==3.2.1
python-slugify==8.0.4
gunicorn==21.2.0
gevent==23.9.1
redis==5.0.1
Faker==19.13.0
